IsPressed & Simplified Draw
Added IsPressed to all input methods and simplified Draw methods so that it's easier to draw things

+++ include/buttons.h
@@ -4,7 +4,7 @@
 #include "api.h"
 #include <cstdint>
 
-enum class KARAKURI_API GamePadButtons : int
+enum class KARAKURI_API GamePadButtons : uint8_t
 {
 	FACE_DOWN = 0,
 	FACE_RIGHT = 1,
